LayoutTests/crypto/workers/subtle/rsa-oaep-import-key-unwrap-key-expected.txt

[Worker] Test unwrapping a JWK oct key with RSA-OAEP using an imported key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/rsa-oaep-import-key-unwrap-key.js
PASS [Worker] unwrappedKey.kty is 'oct'
PASS [Worker] Base64URL.parse(unwrappedKey.k) is rawKey
PASS [Worker] unwrappedKey.alg is 'A128CBC'
PASS [Worker] unwrappedKey.key_ops is ['decrypt', 'encrypt']
PASS [Worker] unwrappedKey.ext is true
PASS successfullyParsed is true

TEST COMPLETE
